Changed STN1110
SCANTOOL changed to STN1110-MM, changed layout of pins on symbol from
QFN package like to logical groups (i think... :P). THIS SYMBOL SHOULD
BE DELETED, but it's not my choice ;)

Also made new element STN1110 with all avaiable packages (QFN-28, SO-28,
DIL28) need to be checked. Now packages QFN-28 are doubled and one
should be deleted, but it's not my choice.