Updated LED-0603
Made these pads slightly wider, and spaced them out a bit more to match
the actual metal pads on the bottom of our 0603 LEDs. I proto'd these
and the fillets were more regular and the LEDs did not rotate to either
side. This will help us avoid "bill-boarding" and lost LEDs during
reflow and/or handling.