TACTILE_SWITCH_SMD.PAC update
Scooted the pads out slightly to allow for better fillet on toe of each
leg. We've proto'd and used this on the new pro mini revision v14, and
it exposes a much better fillet for AOI. Also, it still maintains a good
heal on each leg for strength of connection.

Had to also update TAC_SWITCH.dev (because this package is used there).

Also included a new variant within the SWITCH-MOMENTARY-2.dev for
"REDUNDANT" variant. This links each pin on the symbol to two pads on
the package causing a redundant circuit for the button.