Added NPN Transistor and P-MOSFET variants
Not all TO-92 Transistors have the same pin mappings.
For Example, the 2N3904 is opposite from the BC547 and P2N2222A

This could cause really confusing problems for people using the
Sparkfun Eagle Library if they happen to get a transistor with the wrong pin
map.

Added Pin Mapping Variants for:

TRANSISTOR_NPN
* BC547 - COM-08928 (TO-92 45V 100mA) (1.Collector 2.Base 3.Emitter)
* 2N3904 - COM-00521 (TO-92 40V 200mA) (1.Emitter 2.Base 3.Collector)
* P2N2222A - COM-12852 (TO-92 40V 600mA) (1.Collector 2.Base 3.Emitter)

The Library was also missing the TO-220V variant for the P-Channel Mosfet

Added Pin Mapping Variant for:
MOSFET-PCHANNEL
* FQP27P06 - COM-10349 (TO-220 -60V -27A) (1.Gate 2.Source 3.Drain)

Also added comments and links in the descriptions.