Added locking FP for USB-A-SMD
Did some testing on this footprint to find the best locking foorprint
fit. Turns out that nudging the outside mounting pads just 8 mil forms
the perfect brace. Tight enough to hold position during reflow, but not
so tight that it's really hard to hand place.

Keeping the "non-locking" package and variant in the device, for use
with machine pnp.

-new package: USB-A-SMT-MALE-LOCKING.pac

-also, new variant within USB.dev