Updated SFE_LOGOS to have .0002 width lines
We have had some issues trying to CAM up boards in production that use
polygons that have their outline drawn with a zero-width line. Bumping
the lines up to just 0.0002 inch, causes the error to go away.

I updated all varieties of the SFE logo, including all the sizes of the
"flames alone" and the "name and flames".