Modified/added tKeepout on all SMD resistor packages
All of these packages are now updated to include an accurate outer
rectangle (drawn with a 0.002 inch line) around the outside of each
package on tKeepout layer (39). The lines are drawn so that the outer
edge of the line itself is 0.005 inch away from the outer edge of any
tStop. This is to ensure that any neighboring components should be
placed clear of this tKeepout line and this will form a 0.010 inch space
of mask. This ensures that our fabrication house will actually produce
the line of mask in between tight components. If it is smaller than
0.010 inch, then often times it is smaller than spec and left out.