Added optional attributes to frames
Added two attributes to the document info block: DESIGNER and REV

If left blank (defualt) they will not appear on the schematic, however if you right click the document block and click "attributes" it will allow you to set these fields. Doing so is nicer than just placing text because it stores the meta data in a known location that could be accessed by scripts etc.

This should have no impact on anyone that doesn't want to use it. If not then let me know or just revert the commit ;)