Remove PROD_ID from DIP04 SMD switch
This part (PROD_ID 12418) is a dead prototype with no stand-in components in production. Removing the PROD_ID in the Eagle library retains the history that it was a part (potentially) used in SF designs, but should alert designers of new boards that it is not a standard part that is valid for production (esp. when using the BOM tool)