USB-C connector - removing old 2-layer footprint "USB-C-16P"
The pads version is better. It helps prevents mistakes: If you incorrectly use USB-C-16-2-LAYER-PADS on a 4-layer, it will save you. This is because it will still cause the inner layer polygons to back away from the annular rings on the inner layers of the pads - preventing unwanted electrical connections between inner layers.

Note, it is highly recommended that you use USB-C-16P-4LAYER-ISOLATED for any 4-layer board designs.